feat(service): interface css added for controlling page spread and bckground

+++ server/services/cleanup.cron.js
@@ -24,7 +24,7 @@ cron.schedule('5 8 * * Sun', async () => {
               const EIGHTHOURS = 1000 * 60 * 60 * 8
               const eightHoursAgo = new Date().getTime() - EIGHTHOURS
 
-              if (file <= eightHoursAgo) {
+              if (file !== 'common-stylesheets' && file <= eightHoursAgo) {
                 await fs.remove(path.join(__dirname, '..', 'static', file))
               }
             }
